% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_query.R
\name{DBFactory_closepool}
\alias{DBFactory_closepool}
\title{Close pool connection}
\usage{
DBFactory_closepool(returnerr = T)
}
\arguments{
\item{returnerr}{whether to return error message}
}
\description{
Close pool connection
}
