% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_query.R
\name{DBFactory_disconnect}
\alias{DBFactory_disconnect}
\title{disconnect the connection saved in env}
\usage{
DBFactory_disconnect(type = "sqlserver", returnerr = F)
}
\arguments{
\item{type}{type of connection. sqlserver, mysql, sybaseiq, oracle}
}
\description{
disconnect the connection saved in env
}
