% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_query.R
\name{DBFactory_getsetpool}
\alias{DBFactory_getsetpool}
\title{Get or Set Pool connection}
\usage{
DBFactory_getsetpool(constr)
}
\arguments{
\item{constr}{connection string}
}
\value{
a pool driver
}
\description{
Get or Set Pool connection
}
