% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_query.R
\name{DBFactory_getsetpool_oracle}
\alias{DBFactory_getsetpool_oracle}
\title{Get or Set Pool connection Oracle version}
\usage{
DBFactory_getsetpool_oracle(dbname, username, password)
}
\arguments{
\item{constr}{connection string}
}
\value{
a pool driver
}
\description{
Get or Set Pool connection Oracle version
}
