% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbdef.R
\name{parse_dbobj}
\alias{parse_dbobj}
\title{Parse DB from definition file}
\usage{
parse_dbobj(db = "DEV")
}
\arguments{
\item{db}{database server nickname}
}
\value{
dbobj
}
\description{
parse a string DB name
}
