% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_query.R
\name{run_query}
\alias{run_query}
\title{Main Run query function}
\usage{
run_query(stmt, dbobj = NULL, keepcnxn = F)
}
\arguments{
\item{stmt}{statement to call}

\item{dbobj}{dbobj, default to NULL will use default one}

\item{keepcnxn}{default to F, whether keep cnxn open.}
}
\value{
a result data.frame
}
\description{
main function
}
